requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
dryscrape==1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
soupsieve==1.9.1
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
webkit-server==1.0
xvfbwrapper==0.2.9
